+++ llvm/lib/IR/FPEnv.cpp
@@ -80,27 +80,20 @@ llvm::convertRoundingModeToBundle(RoundingMode UseRounding) {
   std::optional<StringRef> RoundingStr;
   switch (UseRounding) {
   case RoundingMode::Dynamic:
-    RoundingStr = "dynamic";
-    break;
+    return "dynamic";
   case RoundingMode::NearestTiesToEven:
-    RoundingStr = "tonearest";
-    break;
+    return "tonearest";
   case RoundingMode::NearestTiesToAway:
-    RoundingStr = "tonearestaway";
-    break;
+    return "tonearestaway";
   case RoundingMode::TowardNegative:
-    RoundingStr = "downward";
-    break;
+    return "downward";
   case RoundingMode::TowardPositive:
-    RoundingStr = "upward";
-    break;
+    return "upward";
   case RoundingMode::TowardZero:
-    RoundingStr = "towardzero";
-    break;
+    return "towardzero";
   default:
-    break;
+    return std::nullopt;
   }
-  return RoundingStr;
 }
 
 std::optional<fp::ExceptionBehavior>
@@ -143,16 +136,14 @@ llvm::convertExceptionBehaviorToBundle(fp::ExceptionBehavior UseExcept) {
   std::optional<StringRef> ExceptStr;
   switch (UseExcept) {
   case fp::ebStrict:
-    ExceptStr = "strict";
-    break;
+    return "strict";
   case fp::ebIgnore:
-    ExceptStr = "ignore";
-    break;
+    return "ignore";
   case fp::ebMayTrap:
-    ExceptStr = "maytrap";
-    break;
+    return "maytrap";
+  default:
+    return std::nullopt;
   }
-  return ExceptStr;
 }
 
 Intrinsic::ID llvm::getConstrainedIntrinsicID(const Instruction &Instr) {