+++ llvm/unittests/IR/IRBuilderTest.cpp
@@ -274,7 +274,7 @@ TEST_F(IRBuilderTest, ConstrainedFP) {
   // instructions.
   Builder.setDefaultConstrainedRounding(RoundingMode::Dynamic);
   Builder.setDefaultConstrainedExcept(fp::ebStrict);
-  Builder.setIsFPConstrained(true, false);
+  Builder.setIsFPConstrained(true);
   auto Parent = BB->getParent();
   Parent->addFnAttr(Attribute::StrictFP);
 
@@ -442,12 +442,12 @@ TEST_F(IRBuilderTest, ConstrainedFPFunctionCall) {
       Function::Create(FTy, Function::ExternalLinkage, "", M.get());
   BasicBlock *CalleeBB = BasicBlock::Create(Ctx, "", Callee);
   IRBuilder<> CalleeBuilder(CalleeBB);
-  CalleeBuilder.setIsFPConstrained(true);
+  CalleeBuilder.resetModeToStrictFP(true);
   CalleeBuilder.setConstrainedFPFunctionAttr();
   CalleeBuilder.CreateRetVoid();
 
   // Now call the empty constrained FP function.
-  Builder.setIsFPConstrained(true);
+  Builder.resetModeToStrictFP(true);
   Builder.setConstrainedFPFunctionAttr();
   CallInst *FCall = Builder.CreateCall(Callee, {});
 
@@ -571,15 +571,14 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
   F->addFnAttr(Attribute::StrictFP);
 
   IRBuilder<> Builder(BB);
-  Builder.setDefaultConstrainedExcept(fp::ebStrict);
-  Builder.setDefaultConstrainedRounding(RoundingMode::TowardZero);
-  Builder.setIsFPConstrained(true);
 
   GlobalVariable *GVDouble = new GlobalVariable(
       *M, Type::getDoubleTy(Ctx), true, GlobalValue::ExternalLinkage, nullptr);
   Value *FnArg = Builder.CreateLoad(GVDouble->getValueType(), GVDouble);
   Function *Fn = Intrinsic::getOrInsertDeclaration(
       M.get(), Intrinsic::nearbyint, {Type::getDoubleTy(Ctx)});
+
+  Builder.resetModeToStrictFP(true);
 
   // A floating-point operation has side effects in strictfp environment even
   // if it has no FP bundles.
@@ -595,7 +594,8 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());
   }
 
-  // Check call with FP bundles, with default (dynamic) rounding mode
+  // A call may have FP bundles, in this case it has a bundle that specifies
+  // default (dynamic) rounding mode.
   // nearbyint(%x) [ "fp.round" (metadata !"dynamic") ]
   {
     SmallVector<OperandBundleDef, 1> Bundles;
@@ -611,7 +611,7 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());
   }
 
-  // Check call with FP bundles, with specific rounding mode
+  // FP bundle may specify a non-default rounding mode.
   // nearbyint(%x) [ "fp.round" (metadata !"towardzero") ]
   {
     SmallVector<OperandBundleDef, 1> Bundles;
@@ -627,7 +627,7 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());
   }
 
-  // Check call with FP bundles, exception behavior is set to default value.
+  // FP bundles may specify exception behavior, here it is the default value.
   // nearbyint(%x) [ "fp.except" (metadata !"strict") ]
   {
     SmallVector<OperandBundleDef, 1> Bundles;
@@ -643,7 +643,7 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());
   }
 
-  // Check call with FP bundles, exception behavior is set to specific value.
+  // FP bundles may specify exception behavior, including non-default values.
   // nearbyint(%x) [ "fp.except" (metadata !"ignore") ]
   {
     SmallVector<OperandBundleDef, 1> Bundles;
@@ -659,19 +659,56 @@ TEST_F(IRBuilderTest, FPBundlesStrict) {
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());
   }
 
-  // Check call with both FP bundles.
+  // FP bundles may specify both rounding mode and exception behavior, here is
+  // the case of default values.
+  // nearbyint(%x) [ "fp.round" (metadata !"dynamic"),
+  //                 "fp.except" (metadata !"strict") ]
+  {
+    SmallVector<OperandBundleDef, 1> Bundles;
+    llvm::addRoundingBundle(Ctx, Bundles, RoundingMode::Dynamic);
+    llvm::addExceptionBundle(Ctx, Bundles, fp::ebStrict);
+    Value *V = Builder.CreateCall(Fn, {FnArg}, Bundles);
+    auto *I = cast<IntrinsicInst>(V);
+    EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_round).has_value());
+    EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_except).has_value());
+    EXPECT_EQ(Intrinsic::nearbyint, I->getIntrinsicID());
+    EXPECT_EQ(RoundingMode::Dynamic, I->getRoundingMode());
+    EXPECT_EQ(fp::ebStrict, I->getExceptionBehavior());
+    MemoryEffects ME = I->getMemoryEffects();
+    EXPECT_TRUE(ME.doesAccessInaccessibleMem());
+  }
+
+  // FP bundles may specify both rounding mode and exception behavior, here is
+  // the case of non-default values.
   // nearbyint(%x) [ "fp.round" (metadata !"towardzero"),
   //                 "fp.except" (metadata !"ignore") ]
   {
     SmallVector<OperandBundleDef, 1> Bundles;
-    llvm::addRoundingBundle(Ctx, Bundles, RoundingMode::NearestTiesToEven);
+    llvm::addRoundingBundle(Ctx, Bundles, RoundingMode::TowardZero);
     llvm::addExceptionBundle(Ctx, Bundles, fp::ebIgnore);
     Value *V = Builder.CreateCall(Fn, {FnArg}, Bundles);
     auto *I = cast<IntrinsicInst>(V);
     EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_round).has_value());
     EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_except).has_value());
     EXPECT_EQ(Intrinsic::nearbyint, I->getIntrinsicID());
-    EXPECT_EQ(RoundingMode::NearestTiesToEven, I->getRoundingMode());
+    EXPECT_EQ(RoundingMode::TowardZero, I->getRoundingMode());
+    EXPECT_EQ(fp::ebIgnore, I->getExceptionBehavior());
+    MemoryEffects ME = I->getMemoryEffects();
+    EXPECT_TRUE(ME.doesAccessInaccessibleMem());
+  }
+
+  // FP bundles are added when IRBuilder specifies rounding or exception
+  // behavior, non-default for the selected mode.
+  Builder.setDefaultConstrainedExcept(fp::ebIgnore);
+  Builder.setDefaultConstrainedRounding(RoundingMode::TowardZero);
+  Builder.setIsFPConstrained(true);
+  {
+    Value *V = Builder.CreateCall(Fn, {FnArg});
+    auto *I = cast<IntrinsicInst>(V);
+    EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_round).has_value());
+    EXPECT_TRUE(I->getOperandBundle(LLVMContext::OB_fp_except).has_value());
+    EXPECT_EQ(Intrinsic::nearbyint, I->getIntrinsicID());
+    EXPECT_EQ(RoundingMode::TowardZero, I->getRoundingMode());
     EXPECT_EQ(fp::ebIgnore, I->getExceptionBehavior());
     MemoryEffects ME = I->getMemoryEffects();
     EXPECT_TRUE(ME.doesAccessInaccessibleMem());